**Refactor `main` branch: Remove obsolete components, enhance forms, and improve user experience**
- Removed `Home.vue` and its login form as part of cleanup.
- Enhanced `CreateApplication.vue`:
  - Added success and error message handling on form submission.
  - Introduced dynamic course list population by fetching API data.
  - Improved application submission with validation, async handling, and loading states.
  - Disabled "Submit" button during submission process.
- Updated `Dashboard.vue`:
  - Added loading spinner for improved user experience.
  - Enhanced table and card components for better visual feedback.
- Refined application structure and adjusted styles for consistent behavior.
- Updated `app.js` to include `bootstrap.js` for extended functionality.

+++ database/migrations/2025_10_14_100153_create_courses_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('courses', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('school_id');
+            $table->unsignedBigInteger('school_id');
             $table->string('name')->unique();
             $table->string('slug')->unique();
             $table->string('description')->nullable();
